Create user_applications controller
The applications controller was handling requests from employers for
applications to a given job as well as requests from users for all of
their applications to any job. As a result, the applications controller
was becoming difficult to read and had to be handed an unnecessary
number of parameters. Moving the logic for users to view all of their
applications into the new user_applications controller is the first
step in fixing this issue.