Create job_applications route
The applications controller used to handle requests from users to view
all their applications to any job along with requests from employers to
view all applications to a given job. Now it only handles the latter, so
this commit makes the route for employers to view all applications to a
given job a more conventional and cleaner format.