Remove extra logic from application indexes
The applications index view used to be used for showing both all
applications by a given user or all applications to a given job. These
are now being handled by separate views, so this commit removes the
logic that was needed to figure out what kind of request the user was
making.